Single docker container can now be used for prod and reasonable dev environment
