--- Dockerfile.dev
@@ -1,21 +1,0 @@
-FROM python:3
-
-SHELL ["/bin/bash", "-c"]
-WORKDIR /task-api
-
-# postgres
-RUN apt-get update
-RUN apt-get --yes install postgresql
-
-# python app
-COPY requirements.txt requirements.txt
-RUN pip install --no-cache-dir -r requirements.txt
-
-# environment
-ENV APP_SETTINGS development
-ENV FLASK_APP run.py
-ENV SECRET development-secret
-ENV DATABASE_URL postgresql://localhost/tasks_api?user=root\&password=dev-password
-
-EXPOSE 5000
-CMD ["/bin/bash"]
