--- dev/Dockerfile.db
@@ -1,1 +1,0 @@
-FROM postgres
