--- sample.env
@@ -1,5 +1,0 @@
-source env/bin/activate
-export FLASK_APP="run.py"
-export SECRET="development-secret"
-export APP_SETTINGS="development"
-export DATABASE_URL="postgresql://localhost/tasks_api"
