+++ uwsgi.ini
@@ -1,0 +1,3 @@
+[uwsgi]
+module = main
+callable = app
